requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
dj-database-url==2.0.0
python-decouple==3.8
celery==5.2.7
redis==4.5.4
stripe==5.4.0
reportlab==4.0.4
drf-yasg==1.21.5
#psycopg2-binary==2.9.6
gunicorn==20.1.0